Move DocSearch out of core tools and rename to ProjectSearch
